requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
anthropic==0.28.0
networkx==3.2.1
httpx==0.26.0
